Users can login
Users can now login with their email and password combination. If email
or password does not match then login failed and message displayed. No
encryption used for password. Sessions are created however you can’t
tell which user is currently logged in. to be added in the future the
ability to display name on top left corner of blog.